test: Add more high-level methods to `MockChain` (#807)
* feat: High level mockchain methods; re-enable and fix kernel tests
