+++ miden-lib/src/transaction/mod.rs
@@ -249,3 +249,34 @@ impl TransactionKernel {
         Ok(TransactionOutputs { account, output_notes })
     }
 }
+
+#[cfg(feature = "testing")]
+impl TransactionKernel {
+    const KERNEL_TESTING_LIB_BYTES: &'static [u8] =
+        include_bytes!(concat!(env!("OUT_DIR"), "/assets/kernels/kernel_library.masl"));
+
+    pub fn kernel_as_library() -> miden_objects::assembly::Library {
+        miden_objects::assembly::Library::read_from_bytes(Self::KERNEL_TESTING_LIB_BYTES)
+            .expect("failed to deserialize transaction kernel library")
+    }
+
+    /// Contains code to get an instance of the [Assembler] that should be used in tests.
+    ///
+    /// This assembler is similar to the assembler used to assemble the kernel and transactions,
+    /// with the difference that it also includes an extra library on the namespace of `kernel`.
+    /// The `kernel` library is added separately because even though the library (`api.masm`) and
+    /// the kernel binary (`main.masm`) include this code, it is not exposed explicitly. By adding it
+    /// separately, we can expose procedures from `/lib` and test them individually.
+    pub fn assembler_testing() -> Assembler {
+        let source_manager = Arc::new(DefaultSourceManager::default());
+        let kernel_library = Self::kernel_as_library();
+
+        Assembler::with_kernel(source_manager, Self::kernel())
+            .with_library(StdLibrary::default())
+            .expect("failed to load std-lib")
+            .with_library(MidenLib::default())
+            .expect("failed to load miden-lib")
+            .with_library(kernel_library)
+            .expect("failed to load kernel library (/lib)")
+    }
+}
