+++ miden-tx/src/testing/mock_host.rs
@@ -1,7 +1,7 @@
 // MOCK HOST
 // ================================================================================================
 
-use alloc::{string::ToString, sync::Arc};
+use alloc::{rc::Rc, string::ToString, sync::Arc};
 
 use miden_lib::transaction::TransactionEvent;
 use miden_objects::{
@@ -10,11 +10,11 @@ use miden_objects::{
 };
 use vm_processor::{
     AdviceExtractor, AdviceInjector, AdviceInputs, AdviceProvider, AdviceSource, ContextId,
-    ExecutionError, Host, HostResponse, MastForest, MemAdviceProvider, ProcessState,
+    ExecutionError, Host, HostResponse, MastForest, MastForestStore, MemAdviceProvider,
+    ProcessState,
 };
 
-use crate::host::AccountProcedureIndexMap;
-
+use crate::{host::AccountProcedureIndexMap, TransactionMastStore};
 // MOCK HOST
 // ================================================================================================
 
@@ -25,17 +25,23 @@ use crate::host::AccountProcedureIndexMap;
 pub struct MockHost {
     adv_provider: MemAdviceProvider,
     acct_procedure_index_map: AccountProcedureIndexMap,
+    mast_store: Rc<TransactionMastStore>,
 }
 
 impl MockHost {
     /// Returns a new [MockHost] instance with the provided [AdviceInputs].
-    pub fn new(account: AccountStub, advice_inputs: AdviceInputs) -> Self {
+    pub fn new(
+        account: AccountStub,
+        advice_inputs: AdviceInputs,
+        mast_store: Rc<TransactionMastStore>,
+    ) -> Self {
         let adv_provider: MemAdviceProvider = advice_inputs.into();
         let proc_index_map =
             AccountProcedureIndexMap::new(account.code_commitment(), &adv_provider);
         Self {
             adv_provider,
             acct_procedure_index_map: proc_index_map.unwrap(),
+            mast_store,
         }
     }
 
@@ -77,8 +83,8 @@ impl Host for MockHost {
         self.adv_provider.set_advice(process, &injector)
     }
 
-    fn get_mast_forest(&self, _node_digest: &Digest) -> Option<Arc<MastForest>> {
-        todo!("add transaction MAST store")
+    fn get_mast_forest(&self, node_digest: &Digest) -> Option<Arc<MastForest>> {
+        self.mast_store.get(node_digest)
     }
 
     fn on_event<S: ProcessState>(
