+++ miden-tx/src/testing/mod.rs
@@ -3,6 +3,8 @@ pub mod executor;
 pub use mock_host::MockHost;
 mod mock_host;
 
+pub mod mock_chain;
+
 pub use tx_context::{TransactionContext, TransactionContextBuilder};
 mod tx_context;
