+++ miden-tx/src/tests/kernel_tests/test_account.rs
@@ -27,8 +27,8 @@ use crate::{
 pub fn test_set_code_is_not_immediate() {
     let tx_context = TransactionContextBuilder::with_standard_account(ONE).build();
     let code = "
-        use.miden::kernels::tx::prologue
-        use.miden::account
+        use.kernel::prologue
+        use.kernel::account
         begin
             exec.prologue::prepare_transaction
             push.1.2.3.4
@@ -62,9 +62,9 @@ pub fn test_set_code_succeeds() {
 
     let code = format!(
         "
-        use.miden::account
-        use.miden::kernels::tx::prologue
-        use.miden::kernels::tx::epilogue
+        use.kernel::account
+        use.kernel::prologue
+        use.kernel::epilogue
 
         {output_notes_data_procedure}
         begin
@@ -119,8 +119,7 @@ pub fn test_account_type() {
 
             let code = format!(
                 "
-                use.miden::kernels::tx::memory
-                use.miden::kernels::tx::account
+                use.kernel::account
 
                 begin
                     exec.account::{}
@@ -157,7 +156,7 @@ pub fn test_account_type() {
 fn test_validate_id_fails_on_insufficient_ones() {
     let code = format!(
         "
-        use.miden::kernels::tx::account
+        use.kernel::account
 
         begin
             push.{ACCOUNT_ID_INSUFFICIENT_ONES}
@@ -185,7 +184,7 @@ fn test_is_faucet_procedure() {
 
         let code = format!(
             "
-            use.miden::kernels::tx::account
+            use.kernel::account
 
             begin
                 push.{account_id}
@@ -219,15 +218,14 @@ fn test_get_item() {
 
         let code = format!(
             "
-            use.miden::account
-            use.miden::kernels::tx::prologue
+            use.kernel::account
+            use.kernel::prologue
 
             begin
                 exec.prologue::prepare_transaction
-
                 # push the account storage item index
                 push.{item_index}
-
+                
                 # assert the item value is correct
                 exec.account::get_item
                 push.{item_value}
@@ -258,9 +256,9 @@ fn test_set_item() {
 
     let code = format!(
         "
-        use.miden::account
-        use.miden::kernels::tx::memory
-        use.miden::kernels::tx::prologue
+        use.kernel::account
+        use.kernel::memory
+        use.kernel::prologue
 
         begin
             exec.prologue::prepare_transaction
@@ -272,10 +270,13 @@ fn test_set_item() {
 
             # assert old value was empty
             padw assert_eqw
-
+            dropw
             # assert the new item value is properly stored
             exec.memory::get_acct_storage_root
             push.{new_root} assert_eqw
+            dropw dropw
+            dropw dropw
+
         end
         ",
         new_value = prepare_word(&new_item_value),
@@ -298,8 +299,8 @@ fn test_get_storage_data_type() {
 
         let code = format!(
             "
-            use.miden::kernels::tx::account
-            use.miden::kernels::tx::prologue
+            use.kernel::account
+            use.kernel::prologue
 
             begin
                 exec.prologue::prepare_transaction
@@ -349,7 +350,7 @@ fn test_get_map_item() {
         let code = format!(
             "
             use.miden::account
-            use.miden::kernels::tx::prologue
+            use.kernel::prologue
 
             begin
                 exec.prologue::prepare_transaction
@@ -402,7 +403,7 @@ fn test_set_map_item() {
     let code = format!(
         "
         use.miden::account
-        use.miden::kernels::tx::prologue
+        use.kernel::prologue
 
         begin
             exec.prologue::prepare_transaction
@@ -451,7 +452,7 @@ fn test_get_vault_commitment() {
     let code = format!(
         "
         use.miden::account
-        use.miden::kernels::tx::prologue
+        use.kernel::prologue
 
         begin
             exec.prologue::prepare_transaction
@@ -491,8 +492,8 @@ fn test_authenticate_procedure() {
 
         let code = format!(
             "
-            use.miden::kernels::tx::account
-            use.miden::kernels::tx::prologue
+            use.kernel::account
+            use.kernel::prologue
 
             begin
                 exec.prologue::prepare_transaction
