+++ objects/src/testing/account.rs
@@ -6,7 +6,7 @@ use alloc::{
 use core::fmt::Display;
 
 use assembly::Assembler;
-use miden_crypto::merkle::MerkleError;
+use miden_crypto::{dsa::rpo_falcon512::SecretKey, merkle::MerkleError};
 use rand::Rng;
 use vm_core::FieldElement;
 
@@ -107,7 +107,7 @@ impl<T: Rng> AccountBuilder<T> {
     }
 
     /// Build an account using the provided `seed`.
-    pub fn with_seed(
+    pub fn build_with_seed(
         mut self,
         seed: Word,
         assembler: Assembler,
@@ -125,7 +125,7 @@ impl<T: Rng> AccountBuilder<T> {
     /// Build an account using the provided `seed` and `storage`.
     ///
     /// The storage items added to this builder will added on top of `storage`.
-    pub fn with_seed_and_storage(
+    pub fn build_with_seed_and_storage(
         mut self,
         seed: Word,
         mut storage: AccountStorage,
@@ -149,6 +149,18 @@ impl<T: Rng> AccountBuilder<T> {
         let account_code = AccountCode::compile(&self.code, assembler)
             .map_err(AccountBuilderError::AccountError)?;
         Ok(Account::from_parts(account_id, vault, storage, account_code, self.nonce))
+    }
+
+    pub fn build_with_auth(
+        self,
+        assembler: &Assembler,
+    ) -> Result<(Account, Word, SecretKey), AccountBuilderError> {
+        let sec_key = SecretKey::new();
+        let pub_key: Word = sec_key.public_key().into();
+
+        let storage_item = SlotItem::new_value(0, 0, pub_key);
+        let (account, seed) = self.add_storage_item(storage_item).build(assembler.clone())?;
+        Ok((account, seed, sec_key))
     }
 }
