+++ objects/src/testing/account_code.rs
@@ -7,27 +7,28 @@ use crate::accounts::AccountCode;
 const MASTS: [&str; 12] = [
     "0xff06b90f849c4b262cbfbea67042c4ea017ea0e9c558848a951d44b23370bec5",
     "0x8ef0092134469a1330e3c468f57c7f085ce611645d09cc7516c786fefc71d794",
-    "0x8fe219b459e7cf3ced88a5666b3396f31cb97d4729d30b8c20aa9667099846c5",
-    "0xe203ef16c929e5a9253b51d3e71dd182fb3b4b707a5ba5c89bd3e92e587c10b2",
-    "0x92ed8069f0df10267fa4f8c6893c76072264f7b5bf9a6ae2c5dc772cb42f2ad3",
-    "0x7d243d87a57897742a551d98fc166d2cc35ab955ea33c2065db278bf2e6fb91c",
-    "0xe370ca786b01d2b64b97fa2b5300bc54adc1df2d6c8c4945067d3a8fce1f3be2",
-    "0xb6303c8fefc51895d988a5a30272ce97bd2699453f81b2fb8042e0dd55b0bfca",
-    "0x0dbb810e899800e1023ea8921463837484b10dcb4ca1657092245a93c9a08953",
-    "0xc23a46da36a2290941ea32e5ae432a22f501699eb09f6ad0b91bf6c70eaa5e9b",
-    "0x806c992fc7f366314a8723e94d74999d2a663b8291c7d15efd83ddf2b00e8c7d",
-    "0x7cbd0632fba52b8e4003d5cc73376fb5807f2ad5072425563a74f5cf02a56250",
+    "0x86472c62c0f9d2f93369f793abd66127f9cf4a77d4339dcacfb118aba0cf79b6",
+    "0xa5e47b6219605992b497ab85404425da4b88ad58789d86ab09bea9ed0ec12897",
+    "0x56723c7bd5e46ce33f99f256ae1b8f4856600744191f8a18d1c572a925f41ced",
+    "0x6be2be94390361792f5becf18f3e916fa2458c67b809ad144d8ec5fb144ce9c3",
+    "0x0f0447bc4eb9a366d8158274427445fcc169949e4ab9092d45ff55c2a7753e2a",
+    "0x3d77d6c0727fa8c78695123bcd9413e88a5d92e72a60453557fb93dfa575c81a",
+    "0x383067a3ef06a0fad1f11ab7707c67c286db851cc9edece3ea53a76520a014fa",
+    "0x59e26d0f909ce76298f12ba6a6a4120b0cf541622128756eb572fd17dbe8732d",
+    "0xe55e8abaa5a3a8ff89537111b490f22983a7012e65c11ead8478f7a645ba49bd",
+    "0xad0d0d771f4a301c658c61366b4436a4b45b7e317d0f3ae2c76f37e1f8bd63e6",
 ];
 
-pub const ACCOUNT_SEND_ASSET_MAST_ROOT: &str = MASTS[1];
+pub const ACCOUNT_ADD_ASSET_TO_NOTE_MAST_ROOT: &str = MASTS[2];
+pub const ACCOUNT_SEND_ASSET_MAST_ROOT: &str = MASTS[8];
 pub const ACCOUNT_INCR_NONCE_MAST_ROOT: &str = MASTS[4];
-pub const ACCOUNT_SET_ITEM_MAST_ROOT: &str = MASTS[5];
-pub const ACCOUNT_SET_MAP_ITEM_MAST_ROOT: &str = MASTS[6];
-pub const ACCOUNT_SET_CODE_MAST_ROOT: &str = MASTS[7];
-pub const ACCOUNT_ADD_ASSET_TO_NOTE_MAST_ROOT: &str = MASTS[8];
-pub const ACCOUNT_REMOVE_ASSET_MAST_ROOT: &str = MASTS[9];
-pub const ACCOUNT_ACCOUNT_PROCEDURE_1_MAST_ROOT: &str = MASTS[10];
-pub const ACCOUNT_ACCOUNT_PROCEDURE_2_MAST_ROOT: &str = MASTS[11];
+pub const ACCOUNT_SET_ITEM_MAST_ROOT: &str = MASTS[10];
+pub const ACCOUNT_SET_MAP_ITEM_MAST_ROOT: &str = MASTS[11];
+pub const ACCOUNT_SET_CODE_MAST_ROOT: &str = MASTS[9];
+pub const ACCOUNT_REMOVE_ASSET_MAST_ROOT: &str = MASTS[7];
+
+pub const ACCOUNT_ACCOUNT_PROCEDURE_1_MAST_ROOT: &str = MASTS[0];
+pub const ACCOUNT_ACCOUNT_PROCEDURE_2_MAST_ROOT: &str = MASTS[1];
 
 pub const CODE: &str = "
     export.foo
@@ -132,7 +133,6 @@ impl AccountCode {
         ";
 
         let code = AccountCode::compile(account_code, assembler).unwrap();
-
         // Ensures the mast root constants match the latest version of the code.
         //
         // The constants will change if the library code changes, and need to be updated so that the
